requirements.txt
# Both the Unearthed CLI and the Unearthed submission pipeline will attempt to install
# any dependencies listed in this file.
Flask==1.1.1
gunicorn==20.0.4
model-archiver==1.0.3
multi-model-server==1.1.1
numpy==1.19.2
pandas==1.1.3
psutil==5.7.2
python-dateutil==2.8.1
sagemaker-inference==1.2.0
retrying==1.3.3
sagemaker-containers==2.8.6.post2
sagemaker-inference==1.2.0
sagemaker-training
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
gast==0.3.3
grpcio~=1.32.0
h5py~=2.10.0
tensorflow-datasets==4.3.0
efficientnet==1.1.1
albumentations==1.0.0
scikit-learn
bayesian-optimization